crash: move checks to index
